# Faster realignment