@marcelm 2024-11-21
Awesome, thanks for doing some performance work. I think this deserves a changelog entry! Can you add that? Feel free to merge afterwards.